+++ packages/orq-rc/RELEASES.md
@@ -4619,3 +4619,13 @@ Based on:
 - [typescript v3.14.0-rc.62] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.62] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.62 - packages/orq-rc
+
+## 2025-10-28 05:01:05
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.640.0 (2.730.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.63] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.63] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.63 - packages/orq-rc